requirements.txt
fastapi==0.89.1
uvicorn==0.20.0
gunicorn==20.1.0
click==8.1.3
pytest==7.2.1
pytest-cov==4.0.0
pytest-asyncio==0.20.3
requests==2.28.2
httpx==0.23.3
flake8==5.0.4
flake8-docstrings==1.6.0
flake8-import-order==0.18.1
flake8-todo==0.7
black==23.1.0
Sphinx==5.3.0
Pallets-Sphinx-Themes==2.0.2
myst-parser==0.18.1
passlib==1.7.4
python-jose==3.3.0
pymongo==4.3.3
python-multipart==0.0.5
redis==4.5.1
aioredis==2.0.1